--- integrations/mongodb_atlas/src/haystack_integrations/document_stores/mongodb_atlas/errors.py
@@ -1,4 +1,0 @@
-class MongoDBAtlasDocumentStoreError(Exception):
-    """Exception for issues that occur in a MongoDBAtlas document store"""
-
-    pass
